# eyewear_node_typescript/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Eyewear API"
  },
  "basePath": "/api/v1",
  "parameters": {
    "Authrization": {
      "in": "header",
      "name": "authorization",
      "type": "string",
      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Indhc2VlbUB5b3BtYWlsLmNvbSIsInVzZXJJZCI6IjYxNmE4MmQyNzQwZTk5MzY4OTNmZTdiNCIsImlhdCI6MTYzNzY2NzkxMSwiZXhwIjoxNjQwMDg3MTExfQ.sOiVb2KjS_5mfUJSPKi9Iymx4mr8LJb3UElGuZ3999Q",
      "default": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Indhc2VlbUB5b3BtYWlsLmNvbSIsInVzZXJJZCI6IjYxNmE4MmQyNzQwZTk5MzY4OTNmZTdiNCIsImlhdCI6MTYzNzY2NzkxMSwiZXhwIjoxNjQwMDg3MTExfQ.sOiVb2KjS_5mfUJSPKi9Iymx4mr8LJb3UElGuZ3999Q"
    },
    "Timezone": {
      "in": "header",
      "name": "timezone",
      "type": "string",
      "example": "Asia/Calcutta",
      "default": "Asia/Calcutta",
      "required": true
    },
    "AcceptLanguage": {
      "in": "header",
      "name": "Accept-Language",
      "type": "string",
      "example": "en/hi",
      "default": "hi",
      "required": true
    }
  },
...
  "paths": {
...
    "/getcheckoutinfo": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "get": {
        "summary": "get checkout information api",
        "description": "get checkout information API",
        "tags": ["Order apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "type": "string",
            "required": true,
            "example": "616a82d2740e9936893fe7b4",
            "default": "616a82d2740e9936893fe7b4"
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/getCheckoutInfo"
            }
          }
        }
      }
    },
...
    "/checkorder": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "post": {
        "summary": "Check Order",
        "description": "Check Order API",
        "tags": ["Order apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "required": true,
                  "type": "string",
                  "example": "616a82d2740e9936893fe7b4",
                  "default": "616a82d2740e9936893fe7b4"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/checkOrder"
            }
          }
        }
      }
    },
    "/applycoupon": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "post": {
        "summary": "Apply coupon for Order",
        "description": "Apply coupon API",
        "tags": ["Order apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "required": true,
                  "type": "string",
                  "example": "616a82d2740e9936893fe7b4",
                  "default": "616a82d2740e9936893fe7b4"
                },
                "coupon_code": {
                  "required": true,
                  "type": "string",
                  "example": "H9Gf38rJq",
                  "default": "H9Gf38rJq"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/applyCoupon"
            }
          }
        }
      }
    },
    "/placeorder": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "post": {
        "summary": "Place Order",
        "description": "Place Order API",
        "tags": ["Order apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "required": true,
                  "type": "string",
                  "example": "616a82d2740e9936893fe7b4",
                  "default": "616a82d2740e9936893fe7b4"
                },
                "coupon_code": {
                  "type": "string",
                  "default": null
                },
                "name": {
                  "required": true,
                  "type": "string",
                  "example": "Test",
                  "default": "Test"
                },
                "payment_method": {
                  "required": true,
                  "type": "string",
                  "example": "razorpay",
                  "default": "razorpay"
                },
                "card": {
                  "required": true,
                  "type": "number",
                  "example": 4242424242424242,
                  "default": 4242424242424242
                },
                "month": {
                  "required": true,
                  "type": "number",
                  "example": 12,
                  "default": 12
                },
                "year": {
                  "required": true,
                  "type": "number",
                  "example": 2025,
                  "default": 2025
                },
                "cvv": {
                  "required": true,
                  "type": "number",
                  "example": 123,
                  "default": 123
                },
                "notes": {
                  "required": true,
                  "type": "string",
                  "example": "Deliver very fast",
                  "default": "Deliver very fast"
                },
                "address_id": {
                  "required": true,
                  "type": "string",
                  "example": "61742ec7d3ece53de339d5ec",
                  "default": "61742ec7d3ece53de339d5ec"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/placeOrder"
            }
          }
        }
      }
    },
    "/orders": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "get": {
        "summary": "Orders",
        "description": "get user orders by status API",
        "tags": ["Order apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "number",
            "example": 1,
            "default": 1,
            "required": true
          },
          {
            "in": "query",
            "name": "user_id",
            "type": "string",
            "required": true,
            "example": "616a82d2740e9936893fe7b4",
            "default": "616a82d2740e9936893fe7b4"
          },
          {
            "in": "query",
            "name": "status",
            "type": "number",
            "required": true,
            "example": 1,
            "default": 1
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/getOrders"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "getCheckoutInfo": {
      "allOf": [
        {
          "$ref": "#/definitions/getCheckoutInfo"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "checkout info retrived successfullly!"
            }
          }
        }
      ]
    },
...
    "checkOrder": {
      "allOf": [
        {
          "$ref": "#/definitions/checkOrder"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "you can checkout"
            }
          }
        }
      ]
    },
    "applyCoupon": {
      "allOf": [
        {
          "$ref": "#/definitions/applyCoupon"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "coupon applied successfullly!"
            }
          }
        }
      ]
    },
    "placeOrder": {
      "allOf": [
        {
          "$ref": "#/definitions/placeOrder"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "order placed successfullly!"
            }
          }
        }
      ]
    },
    "getOrders": {
      "allOf": [
        {
          "$ref": "#/definitions/getOrders"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "user Orders  retrived successfullly!"
            }
          }
        }
      ]
    },
...
  }
}